# feat: improve performance when a payload contains escape sequence