@goccy 2022-02-14
If you can measure it, could you tell us the benchmark result of #333 by this improvement ?