@orisano 2022-02-14
before:
```
time ./go-json-test
./go-json-test  187.09s user 1.00s system 98% cpu 3:11.41 total
```

after:
```
time ./go-json-test
./go-json-test  0.09s user 0.03s system 24% cpu 0.508 total
```